+++ src/Extension.php
@@ -3,7 +3,7 @@
 /**
  * Title: Jigoshop WordPress pay extension
  * Description:
- * Copyright: Copyright (c) 2005 - 2016
+ * Copyright: Copyright (c) 2005 - 2017
  * Company: Pronamic
  *
  * @author Remco Tolsma
@@ -40,8 +40,10 @@ class Pronamic_WP_Pay_Extensions_Jigoshop_Extension {
 		if ( Pronamic_WP_Pay_Extensions_Jigoshop_Jigoshop::is_active() ) {
 			$slug = self::SLUG;
 
-			add_action( "pronamic_payment_status_update_$slug", array( __CLASS__, 'update_status' ), 10, 2 );
-			add_filter( "pronamic_payment_source_text_$slug",   array( __CLASS__, 'source_text' ), 10, 2 );
+			add_action( 'pronamic_payment_status_update_' . $slug, array( __CLASS__, 'update_status' ), 10, 2 );
+			add_filter( 'pronamic_payment_source_text_' . $slug,   array( __CLASS__, 'source_text' ), 10, 2 );
+			add_filter( 'pronamic_payment_source_description_' . self::SLUG,   array( __CLASS__, 'source_description' ), 10, 2 );
+			add_filter( 'pronamic_payment_source_url_' . self::SLUG,   array( __CLASS__, 'source_url' ), 10, 2 );
 		}
 	}
 
@@ -143,5 +145,23 @@ class Pronamic_WP_Pay_Extensions_Jigoshop_Extension {
 		);
 
 		return $text;
+	}
+
+	/**
+	 * Source description.
+	 */
+	public static function source_description( $description, Pronamic_Pay_Payment $payment ) {
+		$description = __( 'Jigoshop Order', 'pronamic_ideal' );
+
+		return $description;
+	}
+
+	/**
+	 * Source URL.
+	 */
+	public static function source_url( $url, Pronamic_Pay_Payment $payment ) {
+		$url = get_edit_post_link( $payment->get_source_id() );
+
+		return $url;
 	}
 }